requirements.txt
opencv-python
pillow
rich
rich-cli
fastapi
requests
instaloader
loguru
pytest
httpx
pytest-asyncio
python-dotenv